test(js): split multi-module test grab-bags into single-module files
Follow-up to the test-catalog rename. Three files bundled unrelated
modules together, which the rename couldn't fix without moving content.
Resolved data-driven (per-module coverage comparison, deep file vs the
grab-bag block):

- dark-mode-and-smtp-settings → dark-mode.test.js + smtp-settings.test.js
  (the only suites for those two modules — split, not deleted).
- admin-audience-customfields-reregistration → DELETED. All three blocks
  (audience smoke, custom-fields, reregistration) are strict subsets of
  the existing deep suites (grab-only covered lines = 0), so removing
  them changes global line coverage by nothing.
- admin-movesubmissions-recruitment-urlshortener → the url-shortener
  block is a strict subset (dropped); the recruitment and move-submissions
  blocks each cover a guard/bail line the deep suite doesn't, so they are
  preserved as single-module siblings:
  recruitment-admin-fetch.test.js + admin-move-submissions-interception.test.js.

Every test file now exercises exactly one source module. No source
touched, discovery is pure glob. JS line coverage stays 83.95% (≥ 82
floor); the ~0.05% delta vs before is run-to-run v8 async noise (an
unrelated ffc-frontend.js line flips between runs too).

96 files / 1116 tests green; ESLint clean.
