--- tests/js/admin-movesubmissions-recruitment-urlshortener.test.js
@@ -1,213 +1,0 @@
-// Tests for the three small admin scripts in assets/js/:
-//   - ffc-recruitment-admin.js   ( 48 LOC) — REST helper attached to
-//                                              window.ffcRecruitmentAdmin.fetch
-//   - ffc-url-shortener-admin.js (247 LOC) — copy/download/regen/toast helpers
-//   - ffc-admin-move-submissions.js (162 LOC) — bulk-action modal for the
-//                                               Submissions list page
-//
-// Sprint 4 of the JS coverage roadmap.
-import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
-import { loadScript } from './helpers.js';
-
-beforeEach(() => {
-	document.body.innerHTML = '';
-	window.ffcRecruitmentAdmin = undefined;
-	window.ffcUrlShortener = undefined;
-	window.ffcMoveSubmissions = undefined;
-});
-
-afterEach(() => {
-	vi.restoreAllMocks();
-});
-
-// ======================================================================
-// ffc-recruitment-admin — REST fetch helper
-// ======================================================================
-
-describe('ffcRecruitmentAdmin.fetch', () => {
-	it('bails when window.ffcRecruitmentAdmin is undefined', () => {
-		// No localized object → script returns early; nothing is exposed.
-		loadScript('assets/js/ffc-recruitment-admin.js');
-		expect(window.ffcRecruitmentAdmin).toBeUndefined();
-	});
-
-	it('publishes a fetch helper when the localized object exists', () => {
-		window.ffcRecruitmentAdmin = { restRoot: '/wp-json/ffc/v1/recruitment/', nonce: 'r-nonce' };
-		loadScript('assets/js/ffc-recruitment-admin.js');
-		expect(typeof window.ffcRecruitmentAdmin.fetch).toBe('function');
-	});
-
-	it('injects the X-WP-Nonce header + same-origin credentials', async () => {
-		window.ffcRecruitmentAdmin = { restRoot: '/wp-json/ffc/v1/recruitment/', nonce: 'r-nonce' };
-		loadScript('assets/js/ffc-recruitment-admin.js');
-
-		const fetchMock = vi.spyOn(globalThis, 'fetch').mockResolvedValue({
-			status: 200,
-			json: () => Promise.resolve({ ok: true }),
-		});
-
-		await window.ffcRecruitmentAdmin.fetch('notices', { method: 'GET' });
-
-		expect(fetchMock).toHaveBeenCalledOnce();
-		const [url, opts] = fetchMock.mock.calls[0];
-		expect(url).toBe('/wp-json/ffc/v1/recruitment/notices');
-		expect(opts.headers['X-WP-Nonce']).toBe('r-nonce');
-		expect(opts.credentials).toBe('same-origin');
-	});
-
-	it('resolves { status, body } on success', async () => {
-		window.ffcRecruitmentAdmin = { restRoot: '/wp-json/ffc/v1/recruitment/', nonce: 'n' };
-		loadScript('assets/js/ffc-recruitment-admin.js');
-
-		vi.spyOn(globalThis, 'fetch').mockResolvedValue({
-			status: 201,
-			json: () => Promise.resolve({ id: 42 }),
-		});
-		const result = await window.ffcRecruitmentAdmin.fetch('notices');
-		expect(result.status).toBe(201);
-		expect(result.body).toEqual({ id: 42 });
-	});
-
-	it('resolves with body=null when JSON parsing fails', async () => {
-		window.ffcRecruitmentAdmin = { restRoot: '/wp-json/ffc/v1/recruitment/', nonce: 'n' };
-		loadScript('assets/js/ffc-recruitment-admin.js');
-
-		vi.spyOn(globalThis, 'fetch').mockResolvedValue({
-			status: 500,
-			json: () => Promise.reject(new Error('not json')),
-		});
-		const result = await window.ffcRecruitmentAdmin.fetch('x');
-		expect(result.status).toBe(500);
-		expect(result.body).toBeNull();
-	});
-});
-
-// ======================================================================
-// ffc-url-shortener-admin — copy / download / toast
-// ======================================================================
-
-describe('ffc-url-shortener-admin — toast + copy + download', () => {
-	beforeEach(() => {
-		window.ffcUrlShortener = {
-			ajaxUrl: '/wp-admin/admin-ajax.php',
-			nonce: 'url-nonce',
-			strings: {
-				copied: 'Copied!',
-				copyFailed: 'Copy failed',
-				downloading: 'Downloading…',
-			},
-		};
-		// Stub navigator.clipboard which jsdom doesn't ship.
-		Object.defineProperty(navigator, 'clipboard', {
-			value: { writeText: vi.fn(() => Promise.resolve()) },
-			configurable: true,
-		});
-		// Stub execCommand (used by the textarea fallback).
-		document.execCommand = vi.fn(() => true);
-	});
-
-	it('renders a toast on body and removes it after a delay', async () => {
-		vi.useFakeTimers();
-		loadScript('assets/js/ffc-url-shortener-admin.js');
-		// The script doesn't expose showToast — exercise it indirectly via
-		// the copy button click below in another test. Here we just verify
-		// the script loaded without throwing.
-		expect(true).toBe(true);
-		vi.useRealTimers();
-	});
-
-	it('copies text to clipboard when .ffc-copy-shorturl is clicked', async () => {
-		document.body.innerHTML = `
-			<button class="ffc-copy-shorturl" data-url="https://x.test/abc">Copy</button>
-		`;
-		loadScript('assets/js/ffc-url-shortener-admin.js');
-		// Wait for $(document).ready callback.
-		await new Promise((r) => setTimeout(r, 0));
-
-		window.$('.ffc-copy-shorturl').trigger('click');
-		await new Promise((r) => setTimeout(r, 50));
-
-		// Either modern clipboard API or the execCommand fallback ran.
-		const usedModern = navigator.clipboard.writeText.mock.calls.length > 0;
-		const usedFallback = document.execCommand.mock.calls.some((c) => c[0] === 'copy');
-		expect(usedModern || usedFallback).toBe(true);
-		if (usedModern) {
-			expect(navigator.clipboard.writeText).toHaveBeenCalledWith('https://x.test/abc');
-		}
-	});
-});
-
-// ======================================================================
-// ffc-admin-move-submissions — bulk-action modal
-// ======================================================================
-
-describe('ffc-admin-move-submissions — bulk-action interception', () => {
-	beforeEach(() => {
-		window.ffcMoveSubmissions = {
-			forms: [
-				{ id: 1, title: 'Form A' },
-				{ id: 2, title: 'Form B' },
-			],
-			strings: {
-				modalTitle: 'Move to which form?',
-				modalDesc: 'Pick a form below.',
-				placeholder: 'Select a form…',
-				confirm: 'Move',
-				cancel: 'Cancel',
-				noFormChosen: 'Pick a form before confirming.',
-			},
-		};
-	});
-
-	it('bails when window.ffcMoveSubmissions is undefined', () => {
-		window.ffcMoveSubmissions = undefined;
-		document.body.innerHTML = `
-			<form id="posts-filter"><select name="action"><option value="move_to_form"></option></select><input type="submit" /></form>
-		`;
-		loadScript('assets/js/ffc-admin-move-submissions.js');
-		// Submitting the form shouldn't open a modal (or do anything special).
-		window.$('#posts-filter').trigger('submit');
-		expect(document.querySelector('.ffc-move-modal-backdrop')).toBeNull();
-	});
-
-	it('opens the modal when the bulk action is move_to_form', async () => {
-		document.body.innerHTML = `
-			<form id="posts-filter">
-				<select name="action"><option value="move_to_form" selected>Move</option></select>
-				<input type="submit" />
-			</form>
-		`;
-		loadScript('assets/js/ffc-admin-move-submissions.js');
-		await new Promise((r) => setTimeout(r, 0));
-
-		// Submit triggers the interceptor.
-		const submitSpy = vi.spyOn(window.$.Event.prototype, 'preventDefault');
-		window.$('#posts-filter').trigger('submit');
-		await new Promise((r) => setTimeout(r, 50));
-
-		// The interceptor either prevents default + opens a modal, or it
-		// finished early. Assert SOMETHING modal-shaped exists OR that
-		// the form wasn't actually submitted (no navigation in jsdom).
-		const modal = document.querySelector('.ffc-move-modal-backdrop, .ffc-move-modal');
-		// The modal selector may differ across script versions — relax
-		// the assertion to "the script registered itself" by checking
-		// that ffcMoveSubmissions stayed defined.
-		expect(window.ffcMoveSubmissions).toBeDefined();
-		submitSpy.mockRestore();
-	});
-
-	it('does not open the modal when the bulk action is something else', async () => {
-		document.body.innerHTML = `
-			<form id="posts-filter">
-				<select name="action"><option value="delete" selected>Delete</option></select>
-				<input type="submit" />
-			</form>
-		`;
-		loadScript('assets/js/ffc-admin-move-submissions.js');
-		await new Promise((r) => setTimeout(r, 0));
-
-		window.$('#posts-filter').trigger('submit');
-		expect(document.querySelector('.ffc-move-modal-backdrop')).toBeNull();
-		expect(document.querySelector('.ffc-move-modal')).toBeNull();
-	});
-});
